add debug_shell(), use it in non-emergency places
If the user requests rd.break=cmdline, that's not an emergency, it's
just debugging. We don't need to run the emergency hook and the message
shouldn't say "Warning", since nothing's really wrong.

So: add a debug_shell function which doesn't emit a warning, doesn't run
the emergency hook, and won't make the system halt.