dracut-lib: add importlib <lib-name> [for <symbol>...]
This gives us an 'importlib' function that can be used to source
the various libraries (e.g. 'net', 'url', 'img') in the initramfs.
If the library is missing, a warning will be printed.

It won't load a library that has been already loaded in this
interpreter, so you can import multiple times with no problems.

For added assurance and maintainability, you can do something like:

  importlib net for ip_to_var set_ifname

This imports everything in net-lib.sh like usual, but it will print
warnings if it can't find variables or functions (or aliases or
commands) named 'ip_to_var' and 'set_ifname'.